+++ py/jsee/__init__.py
@@ -1,4 +1,4 @@
-from .jsee import generate_schema, serve
+from .jsee import generate_schema, serve, create_app
 from .types import (
   Slider, Text, Radio, Select, MultiSelect, Range, Color,
   Markdown, Html, Code, Image, Table, Svg, File,